% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_txt.R
\name{cat_pages}
\alias{cat_pages}
\title{Concatenate OCR'd pages to a single file}
\usage{
cat_pages(files)
}
\arguments{
\item{files}{A vector of TXT files to be concatenated}
}
\value{
The path to the single, concatenated file
}
\description{
Uses \code{readLines} to build a list so that the form-feed metacharacter
(\code{\\f}) can be used to separate pages in concatenated file. This allows
both text from \link[pdf_text]{pdftools} and from OCR-ing to be handled in
the same way.
}
\examples{
# cat_pages(fil_list)
}
\seealso{
\link{load_text}
}

