% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pdf}
\alias{check_pdf}
\title{Check if file is a PDF}
\usage{
check_pdf(file)
}
\arguments{
\item{file}{Path to a PDF to check for embedding source}
}
\value{
Logical: TRUE if a PDF, FALSE if not
}
\description{
Some <file>.pdf aren't actually PDFs and should be ignored
}
\examples{
\dontrun{
res <- check_pdf("test.pdf")
}
}
\seealso{
\code{pdftools::pdf_info}
}

