% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_txt.R
\name{convert_to_imgs}
\alias{convert_to_imgs}
\title{Convert a file (PDF) to per-page images (PNG)}
\usage{
convert_to_imgs(file, verbose = TRUE)
}
\arguments{
\item{file}{Path to a PDF to be converted to per-page images}
}
\value{
Directory to which the per-page images are saved
}
\description{
Tesseract requires images for OCR. Many analyses benefit from being able to
access pages independently.
}
\examples{
\dontrun{
convert_to_imgs("test.pdf")
}
}

