% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_txt.R
\name{get_sorted_files}
\alias{get_sorted_files}
\title{Return a list of correctly sorted imgs for Tesseract OCR}
\usage{
get_sorted_files(path, ext)
}
\arguments{
\item{path}{Path to the directory containing TXT files from Tesseract}

\item{ext}{The extension of the files (e.g., png) to match}
}
\value{
The (correctly) sorted vector of TXT files
}
\description{
We need the list of PNGs to be sorted correctly, but \code{list.files}
returns a naively sorted list. This means that a document with 10 pages would
return as 'file-0.png file-1.png file-10.png...'; when concatenating
the TXTs, the pages would be concatenated out-of-order. This function returns
the files in the correct order for OCR, and that order is then used for
concatenating the OCR'd pages.
}
\examples{
\dontrun{
get_sorted_files("OCR_tmp/doc1/", "png")

# Returns a vector of files such as c(doc1-0.png, doc1-1.png, doc1-2.png,
# ..., doc1-10.png) rather than c(doc1-0.png, doc1-1.png, doc1-10.png,
# doc1-2.png, ...)
}
}

