% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_text}
\alias{load_text}
\title{Load text extracted from a pdf to a list}
\usage{
load_text(file)
}
\arguments{
\item{file}{Path to the file to be loaded}
}
\value{
A list of one element per page, with a vector of lines per element
}
\description{
\link{pdf_to_txt} writes PDF contents to .txt files rather than loading
them directly. This helper function loads both types of files to a list,
such as one returned by \link[pdf_text]{pdftools}.
}
\examples{
# load_text("doc1234.rda")
# load_text("doc1234.txt")
}

