% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_mgmt.R
\name{make_main_dirs}
\alias{make_main_dirs}
\title{Create main directories expected by \code{pdftext}}
\usage{
make_main_dirs(verbose = FALSE)
}
\arguments{
\item{verbose}{Whether to print message with dirs created}
}
\value{
A named list of created directories
}
\description{
OCR'ing from PDFs requires creating one or more image files to be processed
by \code{tesseract}. Page-by-page OCR is often needed, e.g., documents that
contain a mix of pages with and without an embedded text layer. This function
creates the top-level directories expected by the \link{pdftext} package.
}
\examples{
# make_main_dirs(verbose = TRUE)
}
\seealso{
if any see alsos
}

