% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_txt.R
\name{ocr_pages}
\alias{ocr_pages}
\title{Perform optical character recognition on PNGs.}
\usage{
ocr_pages(pngs, fin_file, verbose = TRUE)
}
\arguments{
\item{pngs}{A listing of the temp PNG directory for a PDF}

\item{fin_file}{Path to the 'final' text file to be written}

\item{verbose}{Whether to print processing messages [TRUE]}
}
\value{
The path to the OCR'd text file
}
\description{
Uses Tesseract (which must be installed and on the $PATH) to perform optical
character recognition (OCR) on a pdf. Uses options()$pdftext.tess_conf to
specify a custom config for tesseract, which can be set with
\link{set_tess_conf}.
}
\examples{
\dontrun{
res <- ocr_pages("test.pdf")
}
}
\seealso{
\code{\link{pdf_to_txt}}
}

