% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_txt.R
\name{ocr_pdf}
\alias{ocr_pdf}
\title{Perform optical character recognition on a PDF}
\usage{
ocr_pdf(file, verbose = TRUE)
}
\arguments{
\item{file}{Path to the PDF from which text will be extracted}

\item{verbose}{Print messages if TRUE; silent if FALSE}
}
\description{
Uses Imagemagick and Tesseract (which must be installed and on the $PATH) to
perform optical character recognition (OCR) on a PDF.
}
\examples{
\dontrun{
res <- ocr_pdf("test.pdf")
}
}
\seealso{
\code{\link{pdf_to_txt}}
}

