% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_txt.R
\name{pdf_to_txt}
\alias{pdf_to_txt}
\title{Extract text from a pdf and write to a txt file}
\usage{
pdf_to_txt(file, thres = 0.2, verbose = TRUE)
}
\arguments{
\item{file}{Path to the PDF from which text will be extracted}

\item{thres}{Threshold number of blank pages to be considered mixed [0.2]}

\item{verbose}{Whether to print processing messages [TRUE]}
}
\value{
Nothing
}
\description{
Extract text from a pdf, which may have a text layer that can be extracted
with \link[pdftools]{pdf_text}; or which may be image-based and needs
to be OCR'd with Tesseract. Both routes end with the extracted text written
to a .txt file with form-feed (\code{\\f}) metacharacters separating pages.
}
\details{
Some PDFs include a mix of pages with and without an embedded text layer.
Getting text from the text layer is preferable to OCR (most of the time), and
to determine which approach to use,
}
\examples{
\dontrun{
res <- pdf_to_txt("test.pdf")
}
}
\seealso{
\link[pdftools]{pdf_text} \code{\link{load_text}}
}

