% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_txt.R
\name{run_unpaper}
\alias{run_unpaper}
\title{Run \code{unpaper} to fix rotation angles}
\usage{
run_unpaper(png_dir)
}
\arguments{
\item{png_dir}{Path to a directory of raw PNGs}
}
\value{
png_dir Path to a directory with 'unpaper'-ed PNGs
}
\description{
A significant number of scanned PDFs are slightly rotated, and even a 1-
degree rotation can substantially reduce the accuracy of OCR. This is a
computationally expensive step, but works much better than not doing it.
}
\examples{
\dontrun{
run_unpaper("IMGs/test")
}
}

