% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_mgmt.R
\name{save_imgs}
\alias{save_imgs}
\title{Save the images directory from options()$pdftext.wkdir}
\usage{
save_imgs(dest)
}
\arguments{
\item{dest}{The path to which the images directory will be saved}
}
\value{
TRUE if copy succeeds, else FALSE
}
\description{
PDFs are split into individual pages for OCR and images from each PDF are
saved to a subdirectory of pdftext.wkdir/IMGs/. This is a helper to copy the
subdirectories and files to a permanent location.
}
\examples{
# save_imgs("~/cur_proj/")
}
\seealso{
\link{save_pages} \link{save_txts}
}

