% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_mgmt.R
\name{save_pages}
\alias{save_pages}
\title{Save the pages directory from tempdir}
\usage{
save_pages(dest)
}
\arguments{
\item{dest}{The path to which the images directory will be saved}
}
\value{
TRUE if copy succeeds, else FALSE
}
\description{
PDFs are split into individual pages for OCR and text files from each PDF are
saved to a subdirectory of tempdir()/PAGEs/. This is a helper to copy the
subdirectories and files to a permanent location.
}
\examples{
# save_imgs("~/cur_proj/")
}
\seealso{
\link{save_pages} \link{save_txts}
}

