% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_mgmt.R
\name{save_txts}
\alias{save_txts}
\title{Save the text directory from tempdir}
\usage{
save_txts(dest)
}
\arguments{
\item{dest}{The path to which the images directory will be saved}
}
\value{
TRUE if copy succeeds, else FALSE
}
\description{
PDFs are split into individual pages for OCR and text files from each PDF are
concatenated to a single file at tempdir()/TXTs/. This is a helper to copy
the TXTs directory (and files) to the user's preferred location.
}
\examples{
# save_txts("~/cur_proj/")
}
\seealso{
\link{save_pages} \link{save_txts}
}

