% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_tess_conf}
\alias{set_tess_conf}
\title{Set a custom \code{tesseract} config for OCR}
\usage{
set_tess_conf(conf)
}
\arguments{
\item{conf}{A string specifying the config file name}
}
\value{
Nothing
}
\description{
\code{tesseract} can be called with custom configs to improve OCR. Adding
such a config automatically is infeasible, but if you want to specify your
own config, such as is found in data(asciimostly), then use this function
to set the option.
}
\examples{
set_tess_conf("asciimostly")
}

