% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_wkdir}
\alias{set_wkdir}
\title{Set the option for the working directory}
\usage{
set_wkdir(wkdir)
}
\arguments{
\item{wkdir}{Absolute path to the working directory (where PDFs/ lives)}
}
\value{
Nothing
}
\description{
\code{pdftext} needs a place to put various files during OCR, e.g., PNGs and
pages from running \code{tesseract}
}
\examples{
set_wkdir("/datadrive/data")
}

