% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_embed.R
\name{test_embed}
\alias{test_embed}
\title{Test if a PDF has embedded text.}
\usage{
test_embed(pdf)
}
\arguments{
\item{pdf}{Path to the PDF from which text will be extracted}
}
\value{
A data.frame of summary info about PDF, incl. whether text is embedded
}
\description{
Test if a PDF has embedded text.
}
\examples{
# res <- test_embed("test.pdf")
}

